+++ db/migrate/20210205113939_remove_daily_reminder_type_from_users.rb
@@ -1,0 +1,5 @@
+class RemoveDailyReminderTypeFromUsers < ActiveRecord::Migration[6.1]
+  def change
+    remove_column :users, :daily_reminder_type, :string
+  end
+end